UI Completed
Added all the confirmation and data savings, fixed all the bugs. The yellow errors when executing  dotnet run  are just because there isn't yet a wa
y for the code to convert if something
is null to a non nullable type. It wo
rks 100% fine right now unless you an
swer a true/false with something else like y or n.